moved FakeResponse to tests/util.py
